chore: remove functions in cogs init
